// src/components/pages/style-selection/style-card-form-values.types.ts
export type StyleCardFormValues = {
  name: string;
  sphere: string;
  goal: string;
  brands: string[];
  styles: string[];
  gender: string;
  age: string;
  totalPrice: string;
  't-shirt': string;
  polo: string;
  shirt: string;
  sweater: string;
  hoodie: string;
  jeans: string;
  trousers: string;
  shorts: string;
  blazer: string;
  jacket: string;
  coat: string;
  shoes: string;
  accessories: string;
  height: string;
  shirtSize: string;
  pantsSize: string;
  jeansSize: string;
  legsSize: string;
  shoesSize: string;
  girths: string;
  peculiarities: string[];
  materials: string[];
  colors: string[];
  email: string;
  phone: string;
  socials: string;
  wishes: string;
  messengers: string[];
  interests: string[];
};